Agentic-Finance-Firm/pyproject.toml

[project]
name = "deep-research-example"
version = "0.1.0"
description = "Deep research agent example using deepagents package"
requires-python = ">=3.11"
dependencies = [
    "langchain-openai>=1.0.2",
    "langchain-anthropic>=1.0.3",
    "langchain_tavily>=0.2.13",
    "pydantic>=2.0.0",
    "rich>=14.0.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.20.0",
    "tavily-python>=0.5.0",
    "httpx>=0.28.1",
    "markdownify>=1.2.0",
    "deepagents>=0.2.6",
    "python-dotenv>=1.0.0",
    "langgraph-cli[inmem]>=0.1.55",
    "langchain-google-genai>=3.1.0",
    "typing-extensions>=4.15.0",
    "langchain>=1.2.6",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-huggingface>=1.2.0",
    "langchain-ollama>=1.0.1",
    "langgraph>=1.0.6",
    "langsmith>=0.6.4",
    "langchain-google-vertexai>=3.2.1",
    "langchain-qdrant>=1.1.0",
    "pymupdf>=1.26.7",
    "pypdf2>=3.0.1",
    "unstructured>=0.18.27",
    "openpyxl>=3.1.5",
    "python-pptx>=1.0.2",
    "python-magic>=0.4.27",
    "docx2txt>=0.9",
    "markitdown[all]>=0.1.4",
    "docutils>=0.22.4",
    "xmltodict>=1.0.2",
    "faiss-cpu>=1.13.2",
    "nltk>=3.9.2",
    "numpy>=2.2.6",
    "tiktoken>=0.12.0",
    "ollama>=0.6.1",
    "openai>=2.15.0",
    "google-genai>=1.60.0",
    "beautifulsoup4>=4.14.3",
    "lxml>=6.0.2",
    "selenium>=4.40.0",
    "duckduckgo-search>=8.1.1",
    "ddgs>=9.10.0",
    "wikipedia>=1.4.0",
    "pymysql>=1.1.2",
    "streamlit>=1.53.0",
    "huggingface-hub>=0.36.0",
    "transformers>=4.57.6",
    "tokenizers>=0.22.2",
    "datasets>=4.5.0",
    "evaluate>=0.4.6",
    "accelerate>=1.12.0",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "ipywidgets>=8.1.8",
    "pytube>=15.0.0",
    "youtube-transcript-api>=1.0.3",
    "email-validator>=2.3.0",
    "textblob>=0.19.0",
    "pymupdf4llm>=0.2.9",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "langgraph-checkpoint-postgres>=3.0.3",
    "psycopg>=3.3.2",
    "psycopg-binary>=3.3.2",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.2.0",
    "hf-xet>=1.2.0",
    "langchain-mcp-adapters>=0.2.1",
    "nest-asyncio>=1.6.0",
    "docling[vlm]>=2.69.1",
    "fastembed>=0.7.3",
    "fastembed-gpu>=0.7.3",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.11.1",
    "ruff>=0.6.1",
]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["scripts"]

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    "UP035",
    "D417",
    "E501",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]

[tool.ruff.lint.pydocstyle]
convention = "google"
